requirements.txt
Django==1.11.3
ipython==6.1.0
psycopg2==2.7.1
djangorestframework==3.6.3
djangorestframework-camel-case==0.2.0
django-cors-headers==2.1.0
dj-database-url==0.4.2
django-hashid-field==1.2.3
django-mail-queue==3.1.2
django-model-utils==3.0.0
django-storages==1.6.5
gunicorn==19.7.1
Pillow==4.2.1
raven==6.2.1
stripe==1.66.0
whitenoise==3.3.1
boto3==1.4.7
django-sendgrid-v5==0.6.8
